requirements.txt
smolagents
openai
anthropic
structlog
langfuse
rich
pyautogui
playwright
pillow
python-dotenv
schedule
mcp
duckduckgo-search
pypdf
